# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Filmes",
        "description": "API desenvolvida para uma plataforma de filmes",
        "contact": {
            "email": "josebrunocampanholi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de testes"
        }
    ],
    "paths": {
        "/api/atores": {
            "get": {
                "tags": [
                    "atores"
                ],
                "summary": "Atores cadastrados no sistema",
                "description": "Retorna os atores cadastrados no sistema paginados em 15",
                "operationId": "indexAtor",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "atores": {
                                            "$ref": "#/components/schemas/Ator"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "atores"
                ],
                "summary": "Cria um ator",
                "description": "Método para criar um novo ator no sistema",
                "operationId": "storeAtor",
                "requestBody": {
                    "description": "Obrigatório nome do autor",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "bruno"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Novo Ator cadastrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro nos parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Informe todos os dados obrigatórios"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/atores/{id}": {
            "get": {
                "tags": [
                    "atores"
                ],
                "summary": "Mostra Ator",
                "description": "Retorna ator pelo ID da url",
                "operationId": "showAtor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do ator",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ator": {
                                            "$ref": "#/components/schemas/Ator"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "atores"
                ],
                "summary": "Atualiza Ator",
                "description": "Atualiza um ator, de acordo com o ID",
                "operationId": "updateAtor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do ator",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Nome do ator",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "bruno"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Ator atualizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro nos parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Informe todos os dados obrigatórios"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "atores"
                ],
                "summary": "Deleta ator",
                "description": "Deleta ator pelo ID da url",
                "operationId": "destroyAtor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do ator",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Sem conteúdo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/atores/{id}/filmes": {
            "get": {
                "tags": [
                    "atores"
                ],
                "summary": "Ator dos Filmes",
                "description": "Retorna todos os filmes que o ator faz parte",
                "operationId": "filmesDoAtor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do ator",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "filmes": {
                                            "$ref": "#/components/schemas/Filme"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/categorias": {
            "get": {
                "tags": [
                    "categorias"
                ],
                "summary": "Categorias cadastradas no sistema",
                "description": "Retorna as categorias cadastrados no sistema paginados em 15",
                "operationId": "indexCategoria",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "categorias": {
                                            "$ref": "#/components/schemas/Categoria"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "categorias"
                ],
                "summary": "Cria uma categoria",
                "description": "Método para criar uma nova categoria no sistema",
                "operationId": "storeAtor",
                "requestBody": {
                    "description": "Obrigatório nome da categoria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Ação"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Nova Categoria cadastrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro nos parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Informe todos os dados obrigatórios"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Ator": {
                "required": [
                    "nome"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "description": "Nome do Ator",
                        "type": "string",
                        "format": "string",
                        "example": "John"
                    },
                    "nacionalidade": {
                        "description": "Nacionalidade do Ator",
                        "type": "string",
                        "format": "string"
                    },
                    "obs": {
                        "description": "Abservações a respeito do Ator",
                        "type": "string",
                        "format": "string"
                    },
                    "created_at": {
                        "description": "Criação inicial timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Última mudança timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
            "Categoria": {
                "required": [
                    "nome"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "description": "Nome da Categoria",
                        "type": "string",
                        "format": "string",
                        "example": "Ação"
                    },
                    "created_at": {
                        "description": "Criação inicial timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Última mudança timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "Filme": {
                "required": [
                    "nome"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "description": "Nome do Filme",
                        "type": "string",
                        "format": "string",
                        "example": "Vingadores"
                    },
                    "created_at": {
                        "description": "Criação inicial timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Última mudança timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Coloque um Token no formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}